Record names of files with geotags for re-use
When running in read-only mode, the names of files that have geotags
to copy are now written to a text file which can be re-read using the
new -f option. This speeds up the actual geotagging portion.

This text file has a simple comma-separated format of path to files,
raw file, jpg.

Some minor refactoring along the way too.